services/ai-orchestrator/pyproject.toml
[project]
name = "ai-orchestrator"
version = "0.1.0"
description = "AI Orchestrator for ChatOps - Step 7: RAG Document Search with pgvector"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    # LangChain for NL -> QueryPlan conversion
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.20",
    # PostgreSQL + pgvector for RAG
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
    # Environment management
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
